requirements.txt
# Runtime
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.21
psycopg2-binary==2.9.7
python-dotenv==1.0.0
gunicorn==21.2.0
honcho==1.1.0
flask-talisman==1.1.0
Flask-Cors==4.0.0
retry2==0.9.5

# Testing
nose==1.3.7
pinocchio==0.4.3
factory-boy==3.3.0
Faker==19.6.2
coverage==7.3.1

# Linting
flake8==6.1.0
pylint==2.17.5

# Werkzeug pin (Flask 2.3 compat)
Werkzeug==2.3.7
